% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_behavior.R
\name{figure_behavior_difficulty}
\alias{figure_behavior_difficulty}
\title{figure_behavior_difficulty}
\usage{
figure_behavior_difficulty()
}
\value{
ggplot object
}
\description{
figure3e: do animals shift towards the flashing side more in difficult trials than medium, easy trials?
1. For each session, get mean percentage choosing high from 9 condition: diffculty (easy, medium, hard) x better_side (low, high, none)
2. For each difficulty, subtract delta_low = low - none, delta_high = high - none, and get delta = mean(delta_low + delta_high)
3. Now we have 3, plot all the points for each animal (LMM for each animal)
}
