% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_behavior.R
\name{figure_behavior_difficulty_schematic}
\alias{figure_behavior_difficulty_schematic}
\title{figure_behavior_difficulty_schematic}
\usage{
figure_behavior_difficulty_schematic()
}
\value{
ggplot object
}
\description{
figure3d: plots a schematic bell curve for easy - medium - hard illustration
}
