% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_behavior.R
\name{figure_behavior_example}
\alias{figure_behavior_example}
\title{figure_behavior_example}
\usage{
figure_behavior_example(animal = 2143, show_sessions = FALSE, n_sessions = 10)
}
\arguments{
\item{animal}{subjid from the \code{good_list}}

\item{show_sessions}{Boolean, if TRUE all the session lines will be shown}

\item{n_sessions}{integer, the number of sessions to draw randomly from this subject's dataset}
}
\value{
ggplot object
}
\description{
figure3a: example psychometric curve from one example animal
}
