% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_behavior.R
\name{figure_behavior_history}
\alias{figure_behavior_history}
\title{figure_behavior_population}
\usage{
figure_behavior_history(animal = 2143)
}
\arguments{
\item{animal}{subjid from the \code{good_list}}
}
\value{
ggplot object
}
\description{
figure3f: plot psychometric curves based on previous trial's outcome
}
