% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_behavior.R
\name{figure_behavior_violin}
\alias{figure_behavior_violin}
\title{figure_behavior_violin}
\usage{
figure_behavior_violin(animal = 2143)
}
\arguments{
\item{animal}{subjid from the \code{good_list}}
}
\value{
ggplot object
}
\description{
figure3b: violin plot showing the distribution of % choosing high by condition across sessions
}
