% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_method.R
\name{figure_method_observation}
\alias{figure_method_observation}
\title{figure_method_observation}
\usage{
figure_method_observation()
}
\value{
ggplot object
}
\description{
figure1c: schematic of the observation distribution p(x|s)
}
