% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_model.R
\name{figure_model_pairs}
\alias{figure_model_pairs}
\title{figure_model_pairs}
\usage{
figure_model_pairs(control_ticks = TRUE)
}
\arguments{
\item{control_ticks}{Boolean, if TRUE the lower and upper limits of ticks are set to specified values}
}
\value{
ggplot object
}
\description{
figure5b: aggregated pairs plot using fits from the mixture-BDT model, summarise the fit for each parameter
}
