% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_model.R
\name{figure_model_prediction}
\alias{figure_model_prediction}
\title{figure_model_prediction}
\usage{
figure_model_prediction(animal = 2143)
}
\arguments{
\item{animal}{subjid from the \code{good_list}}
}
\value{
ggplot object
}
\description{
figure5a: plot prediction ribbons from the mixture-BDT model overlayed with subject binned choice data
}
