% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_model.R
\name{figure_model_sigmoid}
\alias{figure_model_sigmoid}
\title{figure_model_sigmoid}
\usage{
figure_model_sigmoid()
}
\value{
ggplot object
}
\description{
figure6: bar plot showing the change in the four-parameter sigmoid terms in different flash conditions
}
