% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_optimality.R
\name{figure_optimality_bars}
\alias{figure_optimality_bars}
\title{figure_optimality_bars}
\usage{
figure_optimality_bars()
}
\value{
ggplot object
}
\description{
figure7b: how sub-optimal are these animals compared to a fully rational, risk-neutral agent with the fit sigma_p?
}
