% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_optimality.R
\name{figure_optimality_tradeoff}
\alias{figure_optimality_tradeoff}
\title{figure_optimality_tradeoff}
\usage{
figure_optimality_tradeoff()
}
\value{
ggplot object
}
\description{
figure7a: make a heatmap between different levels of rho and sigma, color by normalized reward using the same dataset
}
