% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_task.R
\name{figure_task_ccsd}
\alias{figure_task_ccsd}
\title{figure_task_ccsd}
\usage{
figure_task_ccsd()
}
\value{
ggplot object
}
\description{
figure1c: example of how different sigma_s can affect frequency distribution
}
