% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figureS.R
\name{figureS_bdt_shift}
\alias{figureS_bdt_shift}
\title{figureS_sanity}
\usage{
figureS_bdt_shift(sigma_p = 0.3, rho = 1)
}
\arguments{
\item{sigma_p}{numeric, perceptual noise}

\item{rho}{numeric, utility function exponent}
}
\value{
ggplot object
}
\description{
figureS4: generate synthetic datasets to demonstrate how the mixture-BDT only allows for horizontal shift by changing rho
}
