% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figureS.R
\name{figureS_model_syn}
\alias{figureS_model_syn}
\title{figureS_model_syn}
\usage{
figureS_model_syn(
  param_list1 = list(sigma = 0.2, rho = 1, omega_1 = 0.8, omega_2 = 0.5),
  param_list2 = list(sigma = 0.6, rho = 1, omega_1 = 0.8, omega_2 = 0.5)
)
}
\arguments{
\item{_list1}{sigma, rho, omega_1, omega_2}

\item{_list2}{sigma, rho, omega_1, omega_2}
}
\value{
ggplot object
}
\description{
figureS4: plot how changing parameters change the psychometric curves (takes only 2 sets of parameters)
}
