% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figureS.R
\name{figureS_sigmoid}
\alias{figureS_sigmoid}
\title{figureS_sigmoid}
\usage{
figureS_sigmoid(animal = 2143)
}
\arguments{
\item{animal}{subjid from the \code{good_list}}
}
\value{
ggplot object
}
\description{
figureS5: plots sigmoid prediction overlayed with binned subject choice data
}
